Crate kumo_server_memory

This module contains logic to reason about memory usage, implement memory limits, and some helpers to attempt to release cached memory back to the system when memory is low.

Structs

MemoryLimits

Represents a constraint on memory usage

MemoryUsage

Represents the current memory usage of this process

Functions

get_headroom

Returns the amount of headroom; the number of bytes that can be allocated before we hit the soft limit

get_usage_and_limit

Retrieves the current usage and limits. This is a bit of a murky area as, on Linux, the cgroup reported usage appears to be nonsensical when no limits are configured. So we first obtain the limits from cgroups, and if they are set, we return the usage from cgroups along with it, otherwise we get the ulimit limits and look at the more general usage numbers to go with it.

low_memory

Returns true when we are within 10% if the soft limit

purge_thread_cache

To be called when a thread goes idle; it will flush cached memory out of the thread local cache to be returned/reused elsewhere in the system

setup_memory_limit

Initialize the memory thread to monitor memory usage/limits

subscribe_to_memory_status_changes

Returns a receiver that will notify when memory status changes from OK -> !OK or vice versa.